Add getScaleNotes function and refactor scaleNotes computation in App.vue for improved clarity and functionality
